language/runtime/com/redhat/ceylon/compiler/java/metadata/ValueType.java
package com.redhat.ceylon.compiler.java.metadata;

import java.lang.annotation.ElementType;
import java.lang.annotation.Retention;
import java.lang.annotation.RetentionPolicy;
import java.lang.annotation.Target;

/**
 * Annotation applied to a Java class that is a Ceylon class
 * which adheres to the rules for value types (final class,
 * only abstract super classes, all initializer parameters are
 * shared and immutable and no other attributes exist).
 */
@Retention(RetentionPolicy.RUNTIME)
@Target(ElementType.TYPE)
public @interface ValueType {
}
